requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
decorator==5.1.1
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
django-jazzmin==3.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf_api_logger==1.1.15
future==1.0.0
geocoder==1.38.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
kombu==5.3.7
pillow==10.3.0
prompt_toolkit==3.0.45
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-http-client==3.3.7
PyYAML==6.0.1
ratelim==0.1.6
requests==2.32.3
sendgrid==3.6.5
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
uber-rides==0.6.0
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
zope.interface==6.3
